requirements.txt
streamlit
pandas
plotly_express
pandasql